Added local file storage support
Updated User model - removed redundant fields and added avatar field
Added static_content and static dirs to store User avatars
Made corresponding changes to templates and views
