+++ app/account/models.py
@@ -1,12 +1,19 @@
 from django.contrib.auth.models import AbstractUser
 from django.db import models
+from django.templatetags.static import static
+from utils.helpers import get_upload_to_path
 
 
 class User(AbstractUser):
     email = models.EmailField(unique=True)
-    # TODO two fields are redundant, so they should be deleted during next update of the User model
-    first_name = models.CharField(max_length=50)
-    last_name = models.CharField(max_length=50)
+    avatar = models.ImageField(blank=True, null=True, upload_to=get_upload_to_path)
 
     USERNAME_FIELD = 'email'
     REQUIRED_FIELDS = ['first_name', 'last_name']
+
+    @property
+    def avatar_url(self):
+        if self.avatar:
+            return self.avatar.url
+
+        return static('avatar_default.png')
