+++ app/currency/templates/includes/auth_navbar.html
@@ -5,7 +5,7 @@
         <a href="{% url 'password_reset' %}">Reset Password</a>
     {% else %}
         <a href="{% url 'logout' %}">Logout</a>
-        <a href="{% url 'profile' %}">Profile</a>
+        <a href="{% url 'account:profile' %}">Profile</a>
         <a href="{% url 'password_change' %}">Change Password</a>
     {% endif %}
 </p>
