+++ app/settings/settings.py
@@ -1,6 +1,6 @@
 from configparser import ConfigParser
-from pathlib import Path
 from django.urls import reverse_lazy
+from pathlib import Path
 
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent
@@ -126,6 +126,7 @@ USE_TZ = True
 # https://docs.djangoproject.com/en/4.1/howto/static-files/
 
 STATIC_URL = 'static/'
+STATICFILES_DIRS = [BASE_DIR / 'static']
 
 # Default primary key field type
 # https://docs.djangoproject.com/en/4.1/ref/settings/#default-auto-field
@@ -162,3 +163,6 @@ AUTH_USER_MODEL = 'account.User'
 
 HOST = 'localhost:8000'
 HTTP_SCHEMA = 'http'
+
+MEDIA_URL = 'media/'
+MEDIA_ROOT = BASE_DIR.parent / 'static_content' / 'media'
