+++ static_content/.gitignore
@@ -1,0 +1,5 @@
+# Ignore all
+*
+
+# Except .gitignore
+!.gitignore
